React hygiene - one file per component, default export, filename same as component
