+++ src/app/d2-vendors/Vendors.tsx
@@ -1,30 +1,21 @@
 import { StateParams } from '@uirouter/angularjs';
 import { IScope } from 'angular';
 import {
-  DestinyItemComponentSetOfint32,
-  DestinyVendorComponent,
-  DestinyVendorSaleItemComponent,
   DestinyVendorsResponse,
-  DestinyVendorGroup,
-  DestinyVendorDefinition,
-  BungieMembershipType
+  DestinyVendorGroup
   } from 'bungie-api-ts/destiny2';
 import * as React from 'react';
 import { DestinyAccount } from '../accounts/destiny-account.service';
 import { getVendors as getVendorsApi } from '../bungie-api/destiny2-api';
 import { D2ManifestDefinitions, getDefinitions } from '../destiny2/d2-definitions.service';
-import { BungieImage } from '../dim-ui/bungie-image';
-import Countdown from '../dim-ui/countdown';
 import { D2ManifestService } from '../manifest/manifest-service';
-import VendorItems from './vendor-items';
-import { $state, loadingTracker } from '../ngimport-more';
+import { loadingTracker } from '../ngimport-more';
 import './vendor.scss';
 import { DestinyTrackerServiceType } from '../item-review/destiny-tracker.service';
 import { fetchRatingsForVendors } from './vendor-ratings';
 import { Subscription } from 'rxjs/Subscription';
 import { D2StoreServiceType, D2Store } from '../inventory/store-types';
-import { VendorItem } from './vendor-item';
-import { D2ReviewDataCache } from '../destinyTrackerApi/d2-reviewDataCache';
+import Vendor from './Vendor';
 
 interface Props {
   $scope: IScope;
@@ -42,6 +33,9 @@ interface State {
   ownedItemHashes?: Set<number>;
 }
 
+/**
+ * The "All Vendors" page for D2 that shows all the rotating vendors.
+ */
 export default class Vendors extends React.Component<Props, State> {
   private storesSubscription: Subscription;
 
@@ -163,95 +157,4 @@ function VendorGroup({
       )}
     </>
   );
-}
-
-function Vendor({
-  defs,
-  vendor,
-  itemComponents,
-  sales,
-  trackerService,
-  ownedItemHashes,
-  currencyLookups,
-  account
-}: {
-  defs: D2ManifestDefinitions;
-  vendor: DestinyVendorComponent;
-  itemComponents?: DestinyItemComponentSetOfint32;
-  sales?: {
-    [key: string]: DestinyVendorSaleItemComponent;
-  };
-  trackerService?: DestinyTrackerServiceType;
-  ownedItemHashes?: Set<number>;
-  currencyLookups: {
-    [itemHash: number]: number;
-  };
-  account: DestinyAccount;
-}) {
-  const vendorDef = defs.Vendor.get(vendor.vendorHash);
-  if (!vendorDef) {
-    return null;
-  }
-
-  const destinationDef = defs.Destination.get(vendorDef.locations[vendor.vendorLocationIndex].destinationHash);
-  const placeDef = defs.Place.get(destinationDef.placeHash);
-
-  const placeString = [destinationDef.displayProperties.name, placeDef.displayProperties.name].filter((n) => n.length).join(', ');
-
-  const click = () => $state.go('destiny2.vendor', { id: vendor.vendorHash });
-
-  return (
-    <div className="vendor-char-items">
-      <div className="title">
-        <div className="collapse-handle">
-          <BungieImage src={vendorDef.displayProperties.icon} className="vendor-icon"/>
-          <span onClick={click}>{vendorDef.displayProperties.name}</span>
-          <span className="vendor-location">{placeString}</span>
-        </div>
-        <Countdown endTime={new Date(vendor.nextRefreshDate)}/>
-      </div>
-      <VendorItems
-        defs={defs}
-        vendorDef={vendorDef}
-        vendorItems={getVendorItems(account, defs, vendorDef, trackerService, itemComponents, sales)}
-        trackerService={trackerService}
-        ownedItemHashes={ownedItemHashes}
-        currencyLookups={currencyLookups}
-      />
-    </div>
-  );
-}
-
-export function getVendorItems(
-  account: DestinyAccount,
-  defs: D2ManifestDefinitions,
-  vendorDef: DestinyVendorDefinition,
-  trackerService?: DestinyTrackerServiceType,
-  itemComponents?: DestinyItemComponentSetOfint32,
-  sales?: {
-    [key: string]: DestinyVendorSaleItemComponent;
-  }
-) {
-  const reviewCache: D2ReviewDataCache | undefined = trackerService ? trackerService.getD2ReviewDataCache() : undefined;
-
-  if (sales && itemComponents) {
-    const components = Object.values(sales);
-    return components.map((component) => new VendorItem(
-      defs,
-      vendorDef,
-      vendorDef.itemList[component.vendorItemIndex],
-      reviewCache,
-      component,
-      itemComponents
-    ));
-  } else if (vendorDef.returnWithVendorRequest) {
-    // If the sales should come from the server, don't show anything until we have them
-    return [];
-  } else {
-    return vendorDef.itemList.filter((i) =>
-      !i.exclusivity ||
-      i.exclusivity === BungieMembershipType.All ||
-      i.exclusivity === account.platformType
-    ).map((i) => new VendorItem(defs, vendorDef, i, reviewCache));
-  }
 }
