+++ src/app/progress/progress.module.ts
@@ -1,7 +1,7 @@
 import { StateProvider } from '@uirouter/angularjs';
 import { module } from 'angular';
 import { react2angular } from 'react2angular';
-import { Progress } from './progress';
+import { Progress } from './Progress';
 
 // This is the Destiny 2 "Progress" page with milestones and factions.
 export const progressModule = module('progressModule', [])
